+++ tox.ini
@@ -78,8 +78,7 @@ deps =
     coverage[toml]
     responses==0.20.0
     -r{toxinidir}/requirements.txt
-    # OpenTelemetry is in PYDEPS for tempo instrumentation
-    opentelemetry-exporter-otlp-proto-grpc
+    .[lib_pydeps]
     deepdiff
     httpcore==0.14.7
 commands =
@@ -97,8 +96,7 @@ description = Run integration tests
 deps =
     aiohttp
     deepdiff
-    # OpenTelemetry is in PYDEPS for tempo instrumentation
-    opentelemetry-exporter-otlp-proto-grpc
+    .[lib_pydeps]
     # Libjuju needs to track the juju version
     juju ~= 3.1.0
     lightkube >= 0.11
